minor: improvements in code  minor: added sizing  minor: some other fixes and improvements
